refactor(ai-react): useTransaction→usePlugin, flatten client options, drop subRuns
