+++ packages/ai-react/src/plugin.ts
@@ -1,0 +1,1 @@
+export { usePlugin } from './use-plugin'
